Typing

+++ tests/test_models.py
@@ -6,15 +6,15 @@ import numpy as np
 def test_obs_eq():
     obs1 = Observation(
         data=np.arange(20),
-        available_actions=np.ones(5),
-        state=np.ones(10),
+        available_actions=np.ones(5, dtype=np.float32),
+        state=np.ones(10, dtype=np.float32),
         extras=np.arange(5),
     )
 
     obs2 = Observation(
         data=np.arange(20),
-        available_actions=np.ones(5),
-        state=np.ones(10),
+        available_actions=np.ones(5, dtype=np.float32),
+        state=np.ones(10, dtype=np.float32),
         extras=np.arange(5),
     )
 
@@ -24,15 +24,15 @@ def test_obs_eq():
 def test_obs_eq_extras_none():
     obs1 = Observation(
         data=np.arange(20),
-        available_actions=np.ones(5),
-        state=np.ones(10),
+        available_actions=np.ones(5, dtype=np.float32),
+        state=np.ones(10, dtype=np.float32),
         extras=None,
     )
 
     obs2 = Observation(
         data=np.arange(20),
-        available_actions=np.ones(5),
-        state=np.ones(10),
+        available_actions=np.ones(5, dtype=np.float32),
+        state=np.ones(10, dtype=np.float32),
         extras=None,
     )
 
@@ -42,15 +42,15 @@ def test_obs_eq_extras_none():
 def test_obs_hash():
     obs1 = Observation(
         data=np.arange(20),
-        available_actions=np.ones(5),
-        state=np.ones(10),
+        available_actions=np.ones(5, dtype=np.float32),
+        state=np.ones(10, dtype=np.float32),
         extras=np.arange(5),
     )
 
     obs2 = Observation(
         data=np.arange(20),
-        available_actions=np.ones(5),
-        state=np.ones(10),
+        available_actions=np.ones(5, dtype=np.float32),
+        state=np.ones(10, dtype=np.float32),
         extras=np.arange(5),
     )
 
@@ -61,18 +61,18 @@ def test_transition_eq():
     t1 = Transition(
         obs=Observation(
             data=np.arange(20),
-            available_actions=np.ones(5),
-            state=np.ones(10),
+            available_actions=np.ones(5, dtype=np.float32),
+            state=np.ones(10, dtype=np.float32),
             extras=np.arange(5),
         ),
-        action=np.ones(5),
+        action=np.ones(5, dtype=np.float32),
         reward=1.0,
         done=False,
         info={},
         obs_=Observation(
             data=np.arange(20),
-            available_actions=np.ones(5),
-            state=np.ones(10),
+            available_actions=np.ones(5, dtype=np.float32),
+            state=np.ones(10, dtype=np.float32),
             extras=np.arange(5),
         ),
         truncated=False,
@@ -81,18 +81,18 @@ def test_transition_eq():
     t2 = Transition(
         obs=Observation(
             data=np.arange(20),
-            available_actions=np.ones(5),
-            state=np.ones(10),
+            available_actions=np.ones(5, dtype=np.float32),
+            state=np.ones(10, dtype=np.float32),
             extras=np.arange(5),
         ),
-        action=np.ones(5),
+        action=np.ones(5, dtype=np.float32),
         reward=1.0,
         done=False,
         info={},
         obs_=Observation(
             data=np.arange(20),
-            available_actions=np.ones(5),
-            state=np.ones(10),
+            available_actions=np.ones(5, dtype=np.float32),
+            state=np.ones(10, dtype=np.float32),
             extras=np.arange(5),
         ),
         truncated=False,
@@ -105,18 +105,18 @@ def test_transition_hash():
     t1 = Transition(
         obs=Observation(
             data=np.arange(20),
-            available_actions=np.ones(5),
-            state=np.ones(10),
+            available_actions=np.ones(5, dtype=np.float32),
+            state=np.ones(10, dtype=np.float32),
             extras=np.arange(5),
         ),
-        action=np.ones(5),
+        action=np.ones(5, dtype=np.float32),
         reward=1.0,
         done=False,
         info={},
         obs_=Observation(
             data=np.arange(20),
-            available_actions=np.ones(5),
-            state=np.ones(10),
+            available_actions=np.ones(5, dtype=np.float32),
+            state=np.ones(10, dtype=np.float32),
             extras=np.arange(5),
         ),
         truncated=False,
@@ -125,18 +125,18 @@ def test_transition_hash():
     t2 = Transition(
         obs=Observation(
             data=np.arange(20),
-            available_actions=np.ones(5),
-            state=np.ones(10),
+            available_actions=np.ones(5, dtype=np.float32),
+            state=np.ones(10, dtype=np.float32),
             extras=np.arange(5),
         ),
-        action=np.ones(5),
+        action=np.ones(5, dtype=np.float32),
         reward=1.0,
         done=False,
         info={},
         obs_=Observation(
             data=np.arange(20),
-            available_actions=np.ones(5),
-            state=np.ones(10),
+            available_actions=np.ones(5, dtype=np.float32),
+            state=np.ones(10, dtype=np.float32),
             extras=np.arange(5),
         ),
         truncated=False,
